requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrdict==2.0.1
attrs==21.4.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==19.10b0
bleach==4.1.0
boto3==1.21.13
botocore==1.24.13
braceexpand==0.1.7
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.17.1
editdistance==0.6.0
entrypoints==0.4
executing==0.8.3
fasttext==0.9.2
filelock==3.6.0
fonttools==4.29.1
frozendict==2.3.0
frozenlist==1.3.0
fsspec==2022.2.0
ftfy==6.1.1
future==0.18.2
g2p-en==2.1.0
gdown==4.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.4.0
hydra-core==1.1.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.2
inflect==5.4.0
iniconfig==1.1.1
install==1.3.5
ipadic==1.0.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==4.3.21
jedi==0.18.1
jieba==0.42.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaldi-io==0.9.4
kaldi-python-io==1.2.2
kaldiio==2.17.2
kiwisolver==1.3.2
latexcodec==2.0.1
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.1.0
marshmallow==3.14.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mecab-python3==1.0.3
mistune==0.8.4
mpmath==1.2.1
multidict==6.0.2
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nemo-toolkit==1.7.0
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.8
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
omegaconf==2.1.1
onnx==1.11.0
OpenCC==1.1.3
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
pangu==4.0.6.1
parameterized==0.8.1
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pesq==0.0.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip-api==0.0.28
pipreqs==0.4.11
pluggy==1.0.0
pooch==1.6.0
portalocker==2.4.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyannote.core==4.3
pyannote.database==4.1.3
pyannote.metrics==3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pybind11==2.9.1
pybtex==0.24.0
pybtex-docutils==1.0.1
pycocotools==2.0.4
pycparser==2.21
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.11.2
pynini==2.1.5
pyparsing==3.0.7
pypinyin==0.46.0
pyrsistent==0.18.1
PySocks==1.7.1
pystoi==0.3.3
pytest==7.0.1
pytest-runner==6.0.0
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
rapidfuzz==2.0.5
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
sacrebleu==2.0.0
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.5.6
setproctitle==1.2.2
shellingham==1.4.0
shortuuid==1.0.8
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
sox==1.4.1
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.4.1
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.2.0
sympy==1.9
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.13.2
testpath==0.6.0
threadpoolctl==3.1.0
tokenizers==0.11.6
toml==0.10.2
tomli==2.0.1
torch==1.12.0.dev20220224+cu111
torch-stft==0.1.4
torchaudio==0.12.0.dev20220224+cu111
torchmetrics==0.7.2
torchvision==0.13.0.dev20220224+cu111
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
transformers==4.17.0
typed-ast==1.5.2
typer==0.4.0
typing-extensions==4.1.1
Unidecode==1.3.3
urllib3==1.26.8
wandb==0.12.11
wcwidth==0.2.5
webdataset==0.1.62
webencodings==0.5.1
Werkzeug==2.0.3
wget==3.2
widgetsnbextension==3.5.2
wrapt==1.13.3
yarg==0.1.9
yarl==1.7.2
yaspin==2.1.0
youtokentome==1.0.6
zipp==3.7.0
